+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Improved documentation on Nvidia GPU card detection, and error message if not found. 
 * Updated GitHub action (`go.yaml`) to only change the README.md with the result of the change, if pushing to the
   `main` branch.
+* Updated `Buffer.ToHost()` to use fewer CGO calls, speeding it up.
 
 # v0.4.9 - 2024-11-25